refactor: rename
- database to baomap
- database::test to baomap::readonly_mem
